dermatologist-ai/requirements.txt
gunicorn
flask
torch
torchvision
Pillow
requests
